# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Industry Monitor API (勐腊数据采集)",
    "description": "勐腊行业数据监控与采集系统 API。支持行业数据查询、定时采集、缓存管理、告警监控等功能。",
    "version": "1.0.0",
    "contact": {
      "name": "Admin"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "本地开发环境"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Bearer Token 认证。通过 /api/auth/login 获取 token。"
      }
    },
    "schemas": {
...
      "PanelConfigUpdate": {
        "type": "object",
        "properties": {
          "modules": { "type": "array", "items": {}, "description": "面板模块配置" },
          "layout": { "type": "object", "description": "面板布局配置" }
        }
      },
      "PanelDataFillRequest": {
        "type": "object",
        "required": ["granularity", "startDate", "endDate"],
        "properties": {
          "granularity": {
            "type": "string",
            "enum": ["day", "month", "quarter", "year"],
            "description": "粒度"
          },
          "startDate": { "type": "string", "format": "date", "description": "开始日期 (yyyy-MM-dd)", "example": "2025-01-01" },
          "endDate": { "type": "string", "format": "date", "description": "结束日期 (yyyy-MM-dd)", "example": "2025-01-31" },
          "actions": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": ["high", "hot", "chance", "industryViewV2", "industryTrendRange"]
            },
            "description": "要填充的动作列表，不传默认全部"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "detail": { "type": "string", "description": "错误详情" }
        }
      }
    }
  },
  "security": [],
  "paths": {
...
    "/api/panel/config": {
      "get": {
        "tags": ["面板配置"],
        "summary": "获取面板配置",
        "description": "获取当前行业面板配置（公开接口）",
        "operationId": "getPanelConfig",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "面板配置对象，包含 modules 和 layout"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["面板配置"],
        "summary": "更新面板配置",
        "description": "更新行业面板配置",
        "operationId": "updatePanelConfig",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PanelConfigUpdate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "更新后的配置对象"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/panel/tasks": {
      "get": {
        "tags": ["面板配置"],
        "summary": "获取面板任务列表",
        "description": "列出行业面板相关的可用任务",
        "operationId": "getPanelTasks",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "string", "description": "任务 ID" },
                      "name": { "type": "string", "description": "任务名称" },
                      "description": { "type": "string", "description": "任务描述" }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/panel/tasks/{task_id}/run": {
      "post": {
        "tags": ["面板配置"],
        "summary": "运行面板任务",
        "description": "按 ID 触发面板任务（后台运行）",
        "operationId": "runPanelTask",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "description": "任务 ID",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "task started" },
                    "task_id": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "404": {
            "description": "任务未找到",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/panel/data/fill": {
      "post": {
        "tags": ["面板配置"],
        "summary": "数据补录",
        "description": "提交后台任务，为指定日期范围补录缺失的勐腊数据",
        "operationId": "panelDataFill",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PanelDataFillRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "fill started" },
                    "granularity": { "type": "string" },
                    "startDate": { "type": "string" },
                    "endDate": { "type": "string" },
                    "periodKeyCount": { "type": "integer", "description": "周期 key 总数" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
    "/api/sync-tasks/today": {
      "get": {
        "tags": ["同步任务"],
        "summary": "获取今日同步任务",
        "description": "获取今日的同步任务列表",
        "operationId": "getSyncTasksToday",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "tasks": { "type": "array", "items": {}, "description": "任务列表" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/sync-tasks/{log_id}": {
      "get": {
        "tags": ["同步任务"],
        "summary": "获取同步任务详情",
        "description": "按 ID 获取单个同步任务详情",
        "operationId": "getSyncTaskDetail",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "log_id",
            "in": "path",
            "required": true,
            "description": "同步任务日志 ID",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "同步任务详情"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "404": {
            "description": "任务未找到",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    }
  }
}